requirements.txt
python-dotenv
supabase
groq
python-telegram-bot
agno
matplotlib